+++ Vostok.Hosting/Components/Shutdown/ShutdownFactory.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using Vostok.Commons.Time;
+using Vostok.Hosting.Abstractions;
 using Vostok.Logging.Abstractions;
+using Vostok.Metrics;
 using Vostok.ServiceDiscovery;
 using Vostok.ServiceDiscovery.Abstractions;
 
@@ -14,12 +16,15 @@ namespace Vostok.Hosting.Components.Shutdown
         public static (HostingShutdown hosting, ApplicationShutdown application) Create(
             IServiceBeacon serviceBeacon,
             IServiceLocator serviceLocator,
+            IVostokApplicationIdentity identity,
+            IMetricContext instanceMetrics,
             ILog log,
             int? port,
             IReadOnlyList<CancellationToken> tokens,
             TimeSpan totalTimeout,
             TimeSpan beaconTimeout,
-            bool beaconWaitEnabled)
+            bool beaconWaitEnabled,
+            bool sendAnnotation)
         {
             var hasRealBeacon = serviceBeacon is ServiceBeacon;
             var hasRealLocator = serviceLocator is ServiceLocator;
@@ -39,11 +44,14 @@ namespace Vostok.Hosting.Components.Shutdown
                 applicationShutdown, 
                 serviceBeacon, 
                 serviceLocator,
+                identity,
+                instanceMetrics,
                 log,
                 hostingToken,
                 totalTimeout,
                 beaconTimeout, 
-                beaconWaitEnabled);
+                beaconWaitEnabled,
+                sendAnnotation);
 
             return (hostingShutdown, applicationShutdown);
         }